feat: implement full TES API and persistence.
This commit implements a database abstraction with initial support for
PostgreSQL.

It also replaces the previous implementation of job scheduling with a task
orchestration service that uses a single Kubernetes monitoring request for the
pods it is responsible for. Currently the service is hosted in the same process
as the TES API server and the services communicate via in-memory channels, but
in the future the orchestrator could be moved out into its own Kubernetes
service for horizontal scaling.

Task inputs and outputs have been implemented via `planetary-transporter`, a
binary that is built into a container image and scheduled to move files from
cloud storage into a persistent volume and files out of the persistent volume
to cloud storage. The orchestration service is responsible for scheduling pods
that use the transporter image.

Closes #2.
Closes #3.