fix: include output when azcopy fails.
This commit includes the stdout from `azcopy` when it fails.

Additionally, removed the `-o pipefail` from the command script as it isn't
supported by dash (which is ubuntu's `/bin/sh` symlink).

Also fixed an issue where a task might not transition into a running state when
the first executor failed too quickly and the executor is marked as ignoring
failures.