fix: fix task cancellation.
This commit fixes task cancellation by ensuring pods are not created after a
task has been marked canceled.

Additionally, this implements detecting when pods are deleted outside of the
orchestrator and treats the task as having been preempted.

Also fixed a bug in an error message from the transporter for uploading a file
in a directory output.

Changed various log messages from `info` level to `debug`.