fix: fix `planetary-transporter` not download inputs.
This commit fixes `planetary-transporter` so that it will download inputs for
non-Azure URLs.

It refactors the `download_file` method from the Azure implementation into a
generic implementation that can be called for non-Azure URLs.

Also implements passing a `--transporter-image` to `planetary` which can be
passed in from the chart, allowing `helm install` to override the image name
used internally.