fix: `planetary-transporter` not downloading regular file inputs. (#8)
* fix: fix `planetary-transporter` not download inputs.

This commit fixes `planetary-transporter` so that it will download inputs for
non-Azure URLs.

It refactors the `download_file` method from the Azure implementation into a
generic implementation that can be called for non-Azure URLs.

Also implements passing a `--transporter-image` to `planetary` which can be
passed in from the chart, allowing `helm install` to override the image name
used internally.

* fix: transporter not uploading directories correctly.

The path passed to `cloud::copy` was the directory path, resulting in the copy
uploading the contents of the directory instead of the file.

The fix was to use the entry path as the argument to `cloud::copy`.

* chore: update CHANGELOG.