feat: implement claiming orphaned task pods.
This commit implements orphaning task pods when the orchestrator service shuts
down.

Additionally, the orchestrator service now attempts to periodically claim
orphaned nodes.

To facilitate this, the service takes an advisory lock via the database and
searches for orphaned pods; if it finds any, it claims them.

Upon claiming the pod, the pod state is synchronized with the database and
orchestration resumes for the task.