refactor: refactor `planetary` crate.
This commit refactors the `planetary` crate into multiple crates:

* `planetary-api` - implements the TES API server.
* `planetary-orchestrator` - implements the pod orchestrator.
* `planetary-monitor` - implements the orphaned pod monitor.
* `planetary-server` - shared implementation of an Axum server used by the
  above services.

The directory names of the crates removed the redundant `planetary` prefix too.

The monitor implementation is new and now properly handles orphaned and missing
pods.

For orphaned pods, a request is made to the orchestrator service for an
orchestrator to adopt the pod.

For missing pods, the pod is marked as failed in the database and the TES task
changes to the `SYSTEM_ERROR` state.

With these changes, the API and orchestrator services may scale horizontally.