refactor: refactor `planetary` crate. (#9)
* refactor: create a namespace for the planetary install.

This commit updates the README instructions to put resources relating to
running Planetary into a `planetary` namespace.

Consequently, the namespace for resources relating to the execution of TES
tasks has been renamed from `planetary` to `planetary-tasks`.

* feat: implement claiming orphaned task pods.

This commit implements orphaning task pods when the orchestrator service shuts
down.

Additionally, the orchestrator service now attempts to periodically claim
orphaned nodes.

To facilitate this, the service takes an advisory lock via the database and
searches for orphaned pods; if it finds any, it claims them.

Upon claiming the pod, the pod state is synchronized with the database and
orchestration resumes for the task.

* refactor: refactor `planetary` crate.

This commit refactors the `planetary` crate into multiple crates:

* `planetary-api` - implements the TES API server.
* `planetary-orchestrator` - implements the pod orchestrator.
* `planetary-monitor` - implements the orphaned pod monitor.
* `planetary-server` - shared implementation of an Axum server used by the
  above services.

The directory names of the crates removed the redundant `planetary` prefix too.

The monitor implementation is new and now properly handles orphaned and missing
pods.

For orphaned pods, a request is made to the orchestrator service for an
orchestrator to adopt the pod.

For missing pods, the pod is marked as failed in the database and the TES task
changes to the `SYSTEM_ERROR` state.

With these changes, the API and orchestrator services may scale horizontally.

* chore: fix CI.

* refactor: refactor the `cloud` library.

This commit heavily refactors the `cloud` library to make the future
implementation of S3 and GCS easier.

Additionally, it implements ranged downloads and utilizes `tokio-retry2` for
retrying network operations.

* fix: don't bother with events unless there is a terminal.

Also, fix the event handler such that if it lags and missed certain events,
don't panic on a missing transfer or block state.

* feat: add orchestrator options to specify CPU/memory for transporter pods.

* fix: Azure sometimes returns empty error responses.

For some status code, Azure returns empty responses.

Rather than display a deserialization error, default to displaying the
canonical reason for the status code.

* chore: bump MSRV for let chains.

* Update cloud/src/backend/azure.rs

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>

* Update cloud/src/lib.rs

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>

* Update monitor/src/lib.rs

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>

* chore: code review feedback.

* chore: code review feedback.

---------

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>