feat: add S3 backend to `cloud`. (#11)
* feat: add S3 backend to `cloud`.

This commit adds a S3 backend to `cloud` to allow transferring files to and
from AWS S3.

Also includes a refactoring to configuration.

The `cloud` CLI utility now has options to populating the configuration.

The `planetary` helm chart now creates a secret that is used by the transporter
for passing S3 credentials.

* chore: update CHANGELOGs.