feat: use `cloud-copy` crate.
This commit uses the published `cloud-copy` crate and removes the previously
implemented `cloud` crate.

Additionally, a feature was added to log errors encountered in the monitor and
orchestrator in the database to aid in debugging failures, as pod logs may not
persist long enough to diagnose.

Update the transporter crate to output a summary of what was transferred, which
ends up in the corresponding pod table row.

Pod and PVC cleanup is no longer asynchronous as the event handling is already
done in an asynchronous tokio task.