fix: fix issue with transporter not completing. (#16)
* fix: fix issue with transporter not completing.

A previous commit to `transporter` to update its event handling failed to
either move or drop the events sender before attempting to join on the task
responsible for reading the events; this caused a deadlock because the receiver
would never see the channel as closed.

Additionally, this updates the `tes` and `cloud-copy` crates to latest.

* chore: update CHANGELOG.

* chore: update `cloud-copy` dependency.