refactor: improve throughput of scheduled tasks. (#18)
* refactor: improve throughput of scheduled tasks.

This commit refactors `planetary` such that tasks are now run as a single pod
instead of at least three different pods: an inputs pod, N number of executor
pods, and an outputs pod.

The problem with the previous approach is that each pod needs to be scheduled
independently and, while created sequentially, may be scheduled on different
nodes from previously completed pods of the same task.

When a pod in the task terminates, a request to detach the volume from the node
is sent to the CSI driver and another request is made to attach the volume to
the new node the next pod in the task was scheduled on. This can lead to _a
lot_ of CSI driver requests, including backoffs for errors where the volume
hasn't yet been detached from the previous node.

Additionally, there can be a lot of waiting for the next pod in a task to be
scheduled since a task wasn't executing "atomically".

The change to using a single pod has a lot of benefits to the throughput of
`planetary`, mostly by eliminating a lot of requests to the CSI driver, but
also eliminating a lot of scheduling waits and reducing the number of objects
that Kubernetes keeps track of.

To implement the new design, the orchestrator now creates a single pod with an
init container for downloading inputs, an init container for each executor, and
a main container for the outputs. As executors must be executing sequentially,
we must use an init container for each.

The downside to this approach is that task pods show "initializing" from
Kubernetes perspective when they may actually be running an executor. A task
pod only transitions to "running" when all executors have completed and the
outputs container is running.

Additionally, because there is now only a single pod and the inputs and outputs
containers must contact the orchestrator service, the network policy in place
needs to be updated to allow traffic to the orchestrator. As a result, the
orchestrator's API is now protected via a long API key secret which isn't
available to the executor pods.

The orchestrator now no longer deletes Kubernetes resources; instead, the
monitor now batch deletes objects that are marked for garbage collection at
regular intervals, further reducing the request traffic against the Kubernetes
control plane.

On the database side, the `pods` table has been replaced by a `containers`
table that now only stores information relating to terminated containers.

These changes also include an important fix to set the file permissions of
inputs and outputs so that executors that run as non-root can access them.

* Update README.md

* chore: update README.

* chore: code review feedback.

Use a middleware for auth rather than on each endpoint.

* Update orchestrator/src/main.rs

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>

* chore: code review feedback.

---------

Co-authored-by: Clay McLeod <3411613+claymcleod@users.noreply.github.com>