feat: implement shared key Azure Storage authentication.
This commit updates `cloud-copy` to 0.5.0, which swithces over to using
shared key authentication for Azure Storage from expecting the TES
client to send Azure Storage URLs with SAS tokens applied.

Subsequently, the chart can now be deployed with the Azure Storage
authentication information which is written into a K8s secret.

In the future, we should implement managed identity integration with the
clouds we're running in and authentication storage that way; for now,
users will need to manually configure storage authentication.